Always generate opcode for unbounded sequence
The way bounded sequences were added in this repository is incompatible
with the implementation of the serializer VM in Cyclone DDS proper.

Given that this IDL compiler is end-of-life with the replacement due any
day now, and neither the in-memory representation nor the wire
representation is affected by the presence of a bound, the most sensible
fix is to simply assume that all writers of the type will only publish
sequences that respect the bound and simply drop the information from
the instruction sequence.

Should it turn out to be necessary to add them, then an additional
opcode should be added, at which point the incorrect code can easily be
restored.

Signed-off-by: Erik Boasson <eb@ilities.com>